Simplified table mapping logic.
Merged TableMap<T> and ClassMap<T> to simplify and consolidate the
mapping logic. Updated PropertyMap so that it looks for DataAnnotation
attributes to help determine table name, schema, column name, if the
column is a key and if the column is computed or an identity.
Updated Table{TEntity} so insert and update make better use of the
property mappings.