Simplified logic in NumericExtensions and added a Clamp method.
Fixed a bug (breaking change) in the Between extension method that caused it to include the upper and lower bounds by default when it shouldn't have.

Added a new ConstrainedNumeric<T> type that restricts its value to be withing an upper and lower bound. This is similar to the Ada constraints on a type (http://archive.adaic.com/standards/83rat/html/ratl-04-04.html).